Fixed type to Long

+++ src/main/java/hello/data/model/Person.java
@@ -21,7 +21,7 @@ import java.util.stream.Collectors;
 @Entity
 public class Person {
     private static final Logger LOG = LoggerFactory.getLogger(Person.class);
-    private final IntegerProperty id = new SimpleIntegerProperty(this, "id");
+    private final LongProperty id = new SimpleLongProperty(this, "id");
     private final StringProperty name = new SimpleStringProperty(this, "name");
     private Set<PersonTeam> personTeams = new LinkedHashSet<>();
 
@@ -36,14 +36,14 @@ public class Person {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    public Integer getId() {
+    public Long getId() {
         return id.get();
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id.set(id);
     }
-    public IntegerProperty idProperty() {
+    public LongProperty idProperty() {
         return id;
     }
     @Basic
