+++ src/main/java/hello/data/repository/TeamRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by No3x on 11.08.2017.
  */
-public interface TeamRepository extends CrudRepository<Team, Integer> {
+public interface TeamRepository extends CrudRepository<Team, Long> {
 }
